requirements.txt
# FastAPI サーバー
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# HTTP クライアント (データ取得・アップロード)
aiohttp>=3.8.0,<4.0.0

# Supabase データベース接続
supabase==2.13.0
python-dotenv==1.1.0 